Fixed indexer error: distribution/byteSize field should be in `long` type to support possible larger numbers

+++ magda-int-test/src/test/scala/au/csiro/data61/magda/test/util/Generators.scala
@@ -574,7 +574,7 @@ object Generators {
       license <- someBiasedOption(licenseGen)
       rights <- someBiasedOption(arbitrary[String].map(_.take(50).trim))
       accessURL <- someBiasedOption(arbitrary[String].map(_.take(50).trim))
-      byteSize <- someBiasedOption(arbitrary[Int])
+      byteSize <- someBiasedOption(arbitrary[Long])
       format <- someBiasedOption(formatGen(inputCache))
     } yield
       Distribution(
